processing/pyproject.toml
[build-system]
requires = ["setuptools==76.1.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
sse = "processing.cli:main"
sse_download = "processing.cli:download"

[project]
# INFO: When changing `name`, also change in `utils/get_version_from_setup.py`
name = "sound-scape-explorer"
version = "14-beta.0"
description = "SoundScapeExplorer"
authors = [{ name = "Bamdad Sabbagh", email = "hi@bamdad.fr" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = "~=3.11"

dependencies = [
  "wheel",
  "numpy==2.1.3",
  "pandas==2.2.3",
  "h5py==3.13.0",
  "umap-learn==0.5.7",
  "scikit-learn==1.6.1",
  "scikit-maad==1.4.3",
  "rich==14.0.0",
  "inquirerpy==0.3.4",
  "pydantic==2.11.1",
  "gdown==5.2.0",
  "librosa==0.11.0",
  "tensorflow==2.19.0",
  "kagglehub==0.3.11",
  "tensorflow-hub==0.16.1",
  "keras==3.9.2",
  "tf_keras==2.19.0",
]

[project.optional-dependencies]
dev = [
  "black==25.1.0",
  "pytest==8.3.5",
  "pyright==1.1.398",
  "build==1.2.1",
]
